requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
pydantic
pydantic-settings
python-multipart
sqlalchemy
aiosqlite

# Ingestion (Confluence → Azure AI Search)
requests
beautifulsoup4
azure-search-documents
tiktoken